changes after feedback - single module
