--- architecture-hexagonal/domain/src/main/java/com.baeldung.boot.architecture.hexagonal.domain/User.java
@@ -1,53 +1,0 @@
-package com.baeldung.boot.architecture.hexagonal.domain;
-
-
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.OneToMany;
-import java.util.List;
-
-@Entity
-public class User {
-
-    @Id
-    private Long id;
-    private String firstName;
-    private String lastName;
-
-    @OneToMany(mappedBy = "user")
-    private List<com.baeldung.boot.architecture.hexagonal.domain.Book> books;
-
-    public User() {
-    }
-
-    public User(Long id, String firstName, String lastName) {
-        this.id = id;
-        this.firstName = firstName;
-        this.lastName = lastName;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public String getLastName() {
-        return lastName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-}
