--- architecture-hexagonal/notifier/src/com/baeldung/boot/architecture/hexagonal/notifier/Notifier.java
@@ -1,13 +1,0 @@
-package com.baeldung.boot.architecture.hexagonal.notifier;
-
-/**
- * @author Thanos Floros (thanosfloros@strong-programmer.com)
- */
-public interface Notifier {
-
-    void emailSend();
-
-    void smsSend();
-
-
-}
