--- architecture-hexagonal/rest/src/module-info.java
@@ -1,5 +1,0 @@
-module com.baeldung.rest {
-    requires spring.web;
-    requires com.baeldung.domain;
-    exports com.baeldung.boot.architecture.hexagonal.rest;
-}
