+++ architecture-hexagonal/src/main/java/com/baeldung/boot/architecture/hexagonal/app/service/NotifierService.java
@@ -1,0 +1,11 @@
+package com.baeldung.boot.architecture.hexagonal.app.service;
+
+/**
+ * @author Thanos Floros (thanosfloros@strong-programmer.com)
+ */
+public interface NotifierService {
+
+    void emailSend();
+
+    void smsSend();
+}
