+++ architecture-hexagonal/src/main/java/com/baeldung/boot/architecture/hexagonal/domain/entity/User.java
@@ -1,0 +1,53 @@
+package com.baeldung.boot.architecture.hexagonal.domain.entity;
+
+
+import javax.persistence.Entity;
+import javax.persistence.Id;
+import javax.persistence.OneToMany;
+import java.util.List;
+
+@Entity
+public class User {
+
+    @Id
+    private Long id;
+    private String firstName;
+    private String lastName;
+
+    @OneToMany(mappedBy = "user")
+    private List<Book> books;
+
+    public User() {
+    }
+
+    public User(Long id, String firstName, String lastName) {
+        this.id = id;
+        this.firstName = firstName;
+        this.lastName = lastName;
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+
+}
