+++ architecture-hexagonal/src/main/java/com/baeldung/boot/architecture/hexagonal/domain/repository/UserRepository.java
@@ -1,0 +1,15 @@
+package com.baeldung.boot.architecture.hexagonal.domain.repository;
+
+import com.baeldung.boot.architecture.hexagonal.domain.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+
+/**
+ * @author Thanos Floros (thanosfloros@strong-programmer.com)
+ */
+@Repository
+public interface UserRepository extends JpaRepository<User, String> {
+    List<User> findAll();
+}
